+++ packages/terra-modal-manager/CHANGELOG.md
@@ -4,6 +4,14 @@ Changelog
 Unreleased
 ----------
 
+1.5.0 - (August 8, 2017)
+------------------
+### Added
+* Added fullscreen size option
+
+### Fixed
+* Correcting container class name
+
 1.4.0 - (August 1, 2017)
 ------------------
 ### Changed