+++ packages/terra-table/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+1.5.0 - (August 8, 2017)
+------------------
+### Changed
+* Change data-class to be name spaced
+
 1.4.0 - (August 1, 2017)
 ------------------
 ### Fixed